+++ kube/examples/crd_apply.rs
@@ -37,7 +37,7 @@ pub struct FooStatus {
 async fn main() -> anyhow::Result<()> {
     std::env::set_var("RUST_LOG", "info,kube=info");
     env_logger::init();
-    let client = Client::from(Configuration::inferred().await?);
+    let client = Client::from(Configuration::infer().await?);
     let namespace = std::env::var("NAMESPACE").unwrap_or("default".into());
 
     let ssapply = PatchParams {