+++ kube/examples/event_informer.rs
@@ -12,7 +12,7 @@ use futures::{StreamExt, TryStreamExt};
 async fn main() -> anyhow::Result<()> {
     std::env::set_var("RUST_LOG", "info,kube=debug");
     env_logger::init();
-    let client = Client::from(Configuration::inferred().await?);
+    let client = Client::from(Configuration::infer().await?);
 
     let events = Resource::all::<Event>();
     let lp = ListParams::default();