+++ kube/examples/node_informer.rs
@@ -11,7 +11,7 @@ use kube::{
 async fn main() -> anyhow::Result<()> {
     std::env::set_var("RUST_LOG", "info,node_informer=debug,kube=debug");
     env_logger::init();
-    let client = Client::from(Configuration::inferred().await?);
+    let client = Client::from(Configuration::infer().await?);
     let nodes = Resource::all::<Node>();
     let events: Api<Event> = Api::all(client.clone());