+++ kube/examples/node_reflector.rs
@@ -10,7 +10,7 @@ use kube::{
 async fn main() -> anyhow::Result<()> {
     std::env::set_var("RUST_LOG", "info,kube=debug");
     env_logger::init();
-    let client = Client::from(Configuration::inferred().await?);
+    let client = Client::from(Configuration::infer().await?);
 
     let resource = Resource::all::<Node>();
     let lp = ListParams::default()