+++ kube/examples/pod_informer.rs
@@ -12,7 +12,7 @@ use std::env;
 async fn main() -> anyhow::Result<()> {
     env::set_var("RUST_LOG", "info,kube=debug");
     env_logger::init();
-    let client = Client::from(Configuration::inferred().await?);
+    let client = Client::from(Configuration::infer().await?);
     let namespace = env::var("NAMESPACE").unwrap_or("default".into());
 
     let resource = Resource::namespaced::<Pod>(&namespace);