+++ kube/src/api/crds.rs
@@ -160,7 +160,7 @@ mod test {
         struct FooSpec {
             foo: String,
         };
-        let client = Client::from(Configuration::inferred().await.unwrap());
+        let client = Client::from(Configuration::infer().await.unwrap());
         let r1: Api<Foo> = Api::namespaced(client.clone(), "myns");
 
         let r2: Api<Foo> = CustomResource::kind("Foo")