+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 0.31.0 / 2020-03-XX
 ===================
   * Expose `config::Configuration` at root level
-  * Add `Configuration::inferred` as a recommended constructor
+  * Add `Configuration::infer` as a recommended constructor
   * Rename `client::APIClient` to `client::Client`
   * Expose `client::Client` at root level
-  * Add `Client::inferred` as a recommended constructor
+  * Add `Client::infer` as a recommended constructor
   * `Client` now implements `From<Configuration>`
   * Added comprehensive documentation on `Api`
   * Rename `config::KubeConfigLoader` -> `config::ConfigLoader`